Add fail-if-no-tests command-line option to ConsoleLauncher
When the console launcher found no tests it would succeed. This could
happen by user mistake e.g. by improper use of the selector options. In
that case due to launcher succeding users would miss the
misconfiguration. This option is provided so it can be used to make the
launcher fail when no tests are found, helping user's catch such
misconfigurations.

Fixes: #1298